+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.16.0] - (10.10.2024)
+
+## Changes
+
+- add prohibit functions (power, mode, temperature, all)
+- config.schema update
+- redme update
+- cleanup
+
 ## [0.15.0] - (06.10.2024)
 
 ## Changes
@@ -48,7 +57,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - stability and performance improvements
 - logging refactor
 - cleanup
-
 
 ## [0.10.36] - (14.08.2024)